pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.build.targets.wheel]
packages = ["src/imilia"]

[project]
name = "owkin-imilia"
dynamic = ["version"]
description = "IMILIA: interpretable multiple instance learning for inflammation prediction in IBD from H&E whole slide images"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "Thalyssa Baiocco Rodrigues", email = "thalyssa.baiocco-rodrigues@owkin.com"}, {name = "Antoine Olivier", email = "antoine.olivier@owkin.com"}]
keywords = []
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "gitpython>=3.1.45",
    "joblib>=1.5.2",
    # Add your dependencies here or `uv add <package>` to add them
    # e.g. "pandas>=1.3"
    "loguru>=0.7.3",
    "matplotlib>=3.10.7",
    "scikit-learn>=1.7.2",
    "timm>=1.0.22",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]

[project.scripts]
imilia = "imilia.__main__:main"

[project.urls]
Source = "https://github.com/owkin/imilia"
Issues = "https://github.com/owkin/imilia/issues"

[project.optional-dependencies]
# Add your extra features dependencies here, e.g.
# feature1 = ["pandas>=2.x.x", "numpy>=2.x.x"]

[dependency-groups]
dev = [
    "codespell>=2.2.6",
    "ipykernel>=6.29.3",
    "jupyter>=1.0.0",
    "mypy>=1.8.0",
    "notebook>=7.1.1",
    "pre-commit>=3.5.0",
    "pre-commit-hooks>=4.5.0",
    "ruff>=0.3.1",
    "yamllint>=1.35.1",
]
tests = [
    "pytest>=8.0.2",
    "pytest-cov>=4.1.0",
]

[tool.hatch.version]
path = "src/imilia/__init__.py"

[tool.uv]
default-groups = ["dev", "tests"]

[[tool.uv.index]]
name = "owkin"
url = "https://abstra-058264397262.d.codeartifact.eu-west-1.amazonaws.com/pypi/owkin-pypi/simple/"
explicit = true

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
ignore_missing_imports = true
warn_return_any = true
plugins = []

[tool.pytest.ini_options]
log_cli_level = "INFO"

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
lint.ignore = [
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]
lint.select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]
line-length = 120
target-version = "py310"

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.isort]
known-first-party = ["imilia"]
lines-after-imports = 2

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
]

[tool.codespell]
skip = "*.svelte,*.ts,*.js,*.html,*.yaml,*.yml,*.json"
